Update retrofit

See merge request cmu_grupo131/mysmarthome!21

+++ app/src/main/java/com/example/mysmarthome/retrofit/data_models/plug/PlugRelays.kt
@@ -1,10 +1,12 @@
 package com.example.mysmarthome.retrofit.data_models.plug
 
 data class PlugRelays (
+
     val is_on: Boolean,
     val has_timer: Boolean,
-    val timer_started: Float,
-    val timer_duration: Float,
-    val timer_remaining: Float,
+    val timer_started: Long,
+    val timer_duration: Int,
+    val timer_remaining: Int,
     val overpower: Boolean
+
 ) { }
